# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZHAD0 Protocol API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relayers/{id}:
    get:
      operationId: getRelayer
      tags: [relayers]
      summary: Get a single relayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Relayer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relayer"
        "404":
          description: Not found
...
components:
  schemas:
...
    Relayer:
      type: object
      required:
        - id
        - operatorAddress
        - stakedZhad0
        - status
        - uptimePct
        - intentsRelayed
        - earningsEth
        - region
        - registeredAt
      properties:
        id:
          type: string
        operatorAddress:
          type: string
        stakedZhad0:
          type: string
        status:
          type: string
          enum: [active, inactive, slashed]
        uptimePct:
          type: number
        intentsRelayed:
          type: integer
        earningsEth:
          type: string
        region:
          type: string
        registeredAt:
          type: string
...